+++ CodexLib/Properties/AssemblyInfo.cs
@@ -5,7 +5,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTitle("CodexLib")]
 [assembly: AssemblyProduct("CodexLib")]
 
-[assembly: AssemblyVersion("1.1.1.6")]
-[assembly: AssemblyFileVersion("1.1.1.6")]
+[assembly: AssemblyVersion("1.1.1.7")]
+[assembly: AssemblyFileVersion("1.1.1.7")]
 [assembly: ComVisible(false)]
 [assembly: Guid("aaa3ac08-6a5c-495f-b997-3a3b3bc4fc84")]